refactor(notebook): share runtime state adapters
