+++ src/components/notebook/state/projection-lifecycle.ts
@@ -1,0 +1,16 @@
+import { shouldPreserveBootstrapProjection } from "./bootstrap-preservation";
+import { getCellIdsSnapshot } from "./cell-store";
+
+export function shouldPreserveCurrentNotebookProjection({
+  previousIdentity,
+  nextIdentity,
+}: {
+  previousIdentity: string | null;
+  nextIdentity: string | null;
+}): boolean {
+  return shouldPreserveBootstrapProjection({
+    previousIdentity,
+    nextIdentity,
+    visibleCellCount: getCellIdsSnapshot().length,
+  });
+}
